Backend/TFClub/requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.8.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.3
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.23.1
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
